Cleaned migrations, Added save size functionality for electron views

+++ src/Sidekick.Domain/Views/IViewPreferenceRepository.cs
@@ -5,6 +5,6 @@ namespace Sidekick.Domain.Views
     public interface IViewPreferenceRepository
     {
         Task<ViewPreference> Get(View id);
-        Task SaveSize(View id, double width, double height);
+        Task SaveSize(View id, int width, int height);
     }
 }
